LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_element_haspointercapture-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Element.hasPointerCapture test

Test Description: This test checks if Element.hasPointerCapture returns value correctly
Press black rectangle and do not release
Move your pointer to purple rectangle
Release the pointer
Click purple rectangle

The following pointer types were detected: (none).


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN hasPointerCapture
